refactor(core): move trie db implementations to store crate.
