+++ crates/common/Cargo.toml
@@ -32,8 +32,6 @@ hex-literal.workspace = true
 
 [features]
 default = []
-libmdbx = ["ethrex-trie/libmdbx"]
-redb = ["ethrex-trie/redb"]
 c-kzg = ["dep:c-kzg"]
 
 [lib]
