+++ crates/storage/store/trie_db/redb_multitable.rs
@@ -2,9 +2,9 @@ use std::sync::Arc;
 
 use redb::{Database, MultimapTableDefinition};
 
-use crate::TrieError;
+use ethrex_trie::{TrieDB, TrieError};
 
-use super::{utils::node_hash_to_fixed_size, TrieDB};
+use super::utils::node_hash_to_fixed_size;
 
 const STORAGE_TRIE_NODES_TABLE: MultimapTableDefinition<([u8; 32], [u8; 33]), &[u8]> =
     MultimapTableDefinition::new("StorageTrieNodes");
@@ -25,14 +25,14 @@ impl RedBMultiTableTrieDB {
 
 impl TrieDB for RedBMultiTableTrieDB {
     fn get(&self, key: Vec<u8>) -> Result<Option<Vec<u8>>, TrieError> {
-        let read_txn = self.db.begin_read()?;
-        let table = read_txn.open_multimap_table(STORAGE_TRIE_NODES_TABLE)?;
+        let read_txn = self.db.begin_read().map_err(|e| TrieError::DbError(e.into()))?;
+        let table = read_txn.open_multimap_table(STORAGE_TRIE_NODES_TABLE).map_err(|e| TrieError::DbError(e.into()))?;
 
-        let values = table.get((self.fixed_key, node_hash_to_fixed_size(key)))?;
+        let values = table.get((self.fixed_key, node_hash_to_fixed_size(key))).map_err(|e| TrieError::DbError(e.into()))?;
 
         let mut ret = vec![];
         for value in values {
-            ret.push(value?.value().to_vec());
+            ret.push(value.map_err(|e| TrieError::DbError(e.into()))?.value().to_vec());
         }
 
         let ret_flattened = ret.concat();
@@ -45,25 +45,25 @@ impl TrieDB for RedBMultiTableTrieDB {
     }
 
     fn put(&self, key: Vec<u8>, value: Vec<u8>) -> Result<(), TrieError> {
-        let write_txn = self.db.begin_write()?;
+        let write_txn = self.db.begin_write().map_err(|e| TrieError::DbError(e.into()))?;
         {
-            let mut table = write_txn.open_multimap_table(STORAGE_TRIE_NODES_TABLE)?;
-            table.insert((self.fixed_key, node_hash_to_fixed_size(key)), &*value)?;
+            let mut table = write_txn.open_multimap_table(STORAGE_TRIE_NODES_TABLE).map_err(|e| TrieError::DbError(e.into()))?;
+            table.insert((self.fixed_key, node_hash_to_fixed_size(key)), &*value).map_err(|e| TrieError::DbError(e.into()))?;
         }
-        write_txn.commit()?;
+        write_txn.commit().map_err(|e| TrieError::DbError(e.into()))?;
 
         Ok(())
     }
 
     fn put_batch(&self, key_values: Vec<(Vec<u8>, Vec<u8>)>) -> Result<(), TrieError> {
-        let write_txn = self.db.begin_write()?;
+        let write_txn = self.db.begin_write().map_err(|e| TrieError::DbError(e.into()))?;
         {
-            let mut table = write_txn.open_multimap_table(STORAGE_TRIE_NODES_TABLE)?;
+            let mut table = write_txn.open_multimap_table(STORAGE_TRIE_NODES_TABLE).map_err(|e| TrieError::DbError(e.into()))?;
             for (key, value) in key_values {
-                table.insert((self.fixed_key, node_hash_to_fixed_size(key)), &*value)?;
+                table.insert((self.fixed_key, node_hash_to_fixed_size(key)), &*value).map_err(|e| TrieError::DbError(e.into()))?;
             }
         }
-        write_txn.commit()?;
+        write_txn.commit().map_err(|e| TrieError::DbError(e.into()))?;
 
         Ok(())
     }
