+++ crates/storage/trie/Cargo.toml
@@ -19,12 +19,10 @@ libmdbx = { workspace = true, optional = true }
 smallvec = { version = "1.10.0", features = ["const_generics", "union"] }
 digest = "0.10.6"
 lazy_static.workspace = true
-redb = { workspace = true, optional = true }
 
 [features]
 default = []
 libmdbx = ["dep:libmdbx"]
-redb = ["dep:redb"]
 
 [dev-dependencies]
 hex.workspace = true
