+++ crates/vm/Cargo.toml
@@ -43,7 +43,7 @@ default = []
 l2 = []
 c-kzg = ["revm/c-kzg", "ethrex-levm/c-kzg", "ethrex-common/c-kzg"]
 blst = ["revm/blst"]
-libmdbx = ["ethrex-storage/default", "ethrex-common/libmdbx"]
+libmdbx = ["ethrex-storage/default"]
 
 [profile.test]
 opt-level = 3
